New Rakefile

+++ Rakefile
@@ -1,16 +1,23 @@
-desc "Read and import from CSV"
-task :import do 
-  require 'parse'
-end
-
 desc "Migrate Database"
 task :migrate do
   require 'model'
 end
 
+desc "Read and import from CSV"
+task :import => [:clean, :migrate] do 
+  require 'parse'
+end
+
 desc "Clean DB data"
 task :clean do
-  FileUtils.rm 'data.db'
+  if File.exists? 'data.db'
+    FileUtils.rm 'data.db'
+  end
+end
+
+desc "DB console"
+task :console do
+  sh 'sqlite3 data.db'
 end
 
 task :db_empty => [:clean, :migrate]
